more refactoring
- Created a utils subpackage with multiple modules
- Moved all zarr helper functions in utils.zarr
- Moved all api dependency functions in a new "dependencies" module
- Moved "/info" route from zarr_router to base_router
