+++ tests/test_core.py
@@ -6,10 +6,8 @@ import pytest
 import xarray as xr
 
 import xpublish  # noqa: F401
-from xpublish.routers.zarr import _get_data_chunk
-from xpublish.utils import CostTimer
-
-from .utils import get_zmeta
+from xpublish.utils.cache import CostTimer
+from xpublish.utils.zarr import create_zmetadata, get_data_chunk
 
 
 def test_dask_chunks_become_zarr_chunks():
@@ -17,7 +15,7 @@ def test_dask_chunks_become_zarr_chunks():
     data1 = dask.array.zeros((10, 20, 30), chunks=expected)
     data2 = np.zeros((10, 20, 30))
     ds = xr.Dataset({'foo': (['x', 'y', 'z'], data1), 'bar': (['x', 'y', 'z'], data2)})
-    zmeta = get_zmeta(ds)
+    zmeta = create_zmetadata(ds)
 
     assert zmeta['metadata']['foo/.zarray']['chunks'] == expected
     assert zmeta['metadata']['bar/.zarray']['chunks'] == list(data2.shape)
@@ -30,7 +28,7 @@ def test_invalid_dask_chunks_raise():
     ds = xr.Dataset({'foo': (['x', 'y', 'z'], data)})
 
     with pytest.raises(ValueError) as excinfo:
-        _ = get_zmeta(ds)
+        _ = create_zmetadata(ds)
     excinfo.match(r'Zarr requires uniform chunk sizes .*')
 
 
@@ -40,7 +38,7 @@ def test_invalid_encoding_chunks_with_dask_raise():
     ds = xr.Dataset({'foo': (['x', 'y', 'z'], data)})
     ds['foo'].encoding['chunks'] = [8, 5, 1]
     with pytest.raises(NotImplementedError) as excinfo:
-        _ = get_zmeta(ds)
+        _ = create_zmetadata(ds)
     excinfo.match(r'Specified zarr chunks .*')
 
 
@@ -49,18 +47,18 @@ def test_invalid_encoding_chunks_with_numpy_raise():
     ds = xr.Dataset({'foo': (['x', 'y', 'z'], data)})
     ds['foo'].encoding['chunks'] = [8, 5, 1]
     with pytest.raises(ValueError) as excinfo:
-        _ = get_zmeta(ds)
+        _ = create_zmetadata(ds)
     excinfo.match(r'Encoding chunks do not match inferred.*')
 
 
 def test_get_data_chunk_numpy():
     shape = (2, 5)
     data = np.arange(10).reshape(shape)
-    actual = _get_data_chunk(data, '0.0', shape)
+    actual = get_data_chunk(data, '0.0', shape)
     np.testing.assert_equal(data, actual)
 
     with pytest.raises(ValueError) as excinfo:
-        _ = _get_data_chunk(data, '1.0', shape)
+        _ = get_data_chunk(data, '1.0', shape)
     excinfo.match(r'Invalid chunk_id for numpy array*')
 
 
@@ -68,14 +66,14 @@ def test_get_data_chunk_numpy_edge_chunk():
     # 1d case
     out_shape = (12,)
     data = np.arange(10)
-    actual = _get_data_chunk(data, '0', out_shape)
+    actual = get_data_chunk(data, '0', out_shape)
     assert actual.shape == out_shape
     np.testing.assert_equal(data, actual[:10])
 
     # 2d case
     out_shape = (3, 5)
     data = np.arange(10).reshape((2, 5))
-    actual = _get_data_chunk(data, '0.0', out_shape)
+    actual = get_data_chunk(data, '0.0', out_shape)
     np.testing.assert_equal(data, actual[:2, :])
