+++ xpublish/rest.py
@@ -3,7 +3,8 @@ import uvicorn
 import xarray as xr
 from fastapi import FastAPI
 
-from .routers import base_router, common_router, get_cache, get_dataset, zarr_router
+from .dependencies import get_cache, get_dataset
+from .routers import base_router, common_router, zarr_router
 
 
 @xr.register_dataset_accessor('rest')
