+++ xpublish/routers/__init__.py
@@ -1,3 +1,3 @@
-from .base import base_router, get_cache, get_dataset
+from .base import base_router
 from .common import common_router
 from .zarr import zarr_router
